Hotels: expose only one standalone package
Hopefully, there is no longer direct usage of core in hotels or hotels
in the core. It should be therefore standalone (not completely verified
yet).
